Add files via upload
Starter code for navigation of the robot. Probably not where we want to be at, but provides a good starting point for the navigation part of the robot (and maybe if we want to do navigation in JavaScript, a good inspiration to look into), Could have elaborated more on the code such as navigation to the rooms. Please contact Angela if you have any suggestions on how to further elaborate the code to meet the primary goals.